requirements.txt
lark-parser>=0.12.0
syncode>=0.1.0
accelerate>=0.20.0
